Move list chunking method out of scraper into utils.list_utils
