+++ src/auto_xkcd/packages/xkcd_comic/comic/scraper.py
@@ -23,44 +23,6 @@ import httpx
 import hishel
 
 
-def make_list_chunks(
-    input_list: list[t.Any] = None, max_list_size: int = 50
-) -> list[list[t.Any]] | list[t.Any]:
-    """Break a list into smaller lists/"chunks" based on max_list_size.
-
-    Params:
-        input_list (list): The input list to be chunked.
-        max_list_size (int): The maximum size of each chunk.
-
-    Returns:
-        list of lists: List of smaller lists or chunks.
-
-    """
-    assert input_list, ValueError("Missing input list to break into smaller chunks")
-    assert isinstance(input_list, list), TypeError(
-        f"input_list must be a list. Got type: ({type(input_list)})"
-    )
-
-    if max_list_size == 0 or max_list_size is None:
-        log.warning(
-            f"No limit set on list size. Returning full list embedded in another list"
-        )
-
-        return input_list
-
-    chunked_list: list[list] = []
-    ## Loop over objects in list, breaking into smaller list chunks each time
-    #  the iterator reaches max_list_size
-    for i in range(0, len(input_list), max_list_size):
-        chunked_list.append(input_list[i : i + max_list_size])
-
-    log.debug(
-        f"Created [{len(chunked_list)}] list(s) of {max_list_size} or less items."
-    )
-
-    return chunked_list
-
-
 def scrape_missing_comics(
     cache_transport: hishel.CacheTransport = None,
     request_sleep: int = 5,
@@ -85,7 +47,7 @@ def scrape_missing_comics(
         log.warning(
             f"Number of missing comics ({len(missing_comic_nums)}) is greater than the configured max_list_size ({max_list_size}). Breaking into smaller chunks"
         )
-        missing_nums_lists: list[list[int]] = make_list_chunks(
+        missing_nums_lists: list[list[int]] = list_utils.make_list_chunks(
             input_list=missing_comic_nums, max_list_size=max_list_size
         )
 
@@ -160,6 +122,7 @@ def scrape_missing_comics(
             )
 
             return joined_list
+
         except Exception as exc:
             msg = Exception(
                 f"Unhandled exception joining chunked lists of XKCDComic objects into single list. Details: {exc}"
