+++ src/auto_xkcd/utils/list_utils/__init__.py
@@ -1,1 +1,1 @@
-from .methods import join_list_of_lists
+from .methods import join_list_of_lists, make_list_chunks
